Установил className и начал разбираться с css-modules
