+++ .lintstagedrc.json
@@ -1,5 +1,3 @@
 {
-  "*.html": "prettier --write",
-  "*.css": "prettier --write",
-  "*.js": "prettier --write"
+  "*.{html,js,jsx,ts,tsx,css,scss,md}": "prettier --write"
 }
