+++ src/shared/Button/Button.module.css
@@ -1,6 +1,6 @@
 .buttons {
   width: 36rem;
-  height: 23.5rem;
+  /*height: 23.5rem;*/
   display: grid;
   grid-template-columns: 1fr 1fr 1fr;
   grid-gap: 3rem;
