+++ src/shared/index.js
@@ -1,0 +1,7 @@
+export { Buttons } from "./Button/Button";
+export { Checkboxes } from "./Checkbox/Checkbox";
+export { Radios } from "./Radio/Radio";
+export { Inputs } from "./Input/Input";
+export { Searchbars } from "./Searchbar/Searchbar";
+export { PageHeaders } from "./PageHeader/PageHeader";
+export { Dropdowns } from "./Dropdown/Dropdown";
